Checkpoint: infinite loop maybe issue with pagination?
